refactor: StatisticsMachine 클래스 생성 및 메서드 분리
