+++ src/LottoGame.js
@@ -1,14 +1,14 @@
 const { Console } = require("@woowacourse/mission-utils");
 const { REQUEST_MESSAGE } = require("./constants/message.js");
-const { LOTTO_PRIZE_MATCH_COUNT, LOTTO_PRIZE_MONEY } = require("./constants/condition.js");
 const LottoGameView = require("./LottoGameView.js");
-const Validation = require("./Validation.js");
 const LottoPerchaseMachine = require("./LottoPurchaseMachine.js");
 const WinningLotto = require("./WinningLotto.js");
 const LottoDrawMachine = require("./LottoDrawMachine.js");
+const StatisticsMachine = require("./StatisticsMachine.js");
 
 class LottoGame {
   lottos;
+  purchaseAmount;
   winningNumbers;
   bonusNumber;
 
@@ -17,6 +17,7 @@ class LottoGame {
     this.lottoPerchaseMachine = new LottoPerchaseMachine();
     this.winningLotto = new WinningLotto();
     this.lottoDrawMachine = new LottoDrawMachine();
+    this.statisticsMachine = new StatisticsMachine();
   }
 
   play() {
@@ -25,7 +26,8 @@ class LottoGame {
 
   purchaseLottoPhase() {
     this.LottoGameView.requestInput(REQUEST_MESSAGE.PURCHASE_AMOUNT, (purchaseAmount) => {
-      this.lottoPerchaseMachine.insertMoney(purchaseAmount);
+      this.purchaseAmount = purchaseAmount;
+      this.lottoPerchaseMachine.insertMoney(this.purchaseAmount);
       this.lottos = this.lottoPerchaseMachine.purchaseLottos();
 
       this.LottoGameView.printLottoQuantity(this.lottos.length);
@@ -61,80 +63,21 @@ class LottoGame {
       this.bonusNumber
     );
 
-    const prizeStatistics = this.getPrizeStatistics(eachLottoPrize);
-    const totalPrizeMoney = this.getTotalPrizeMoney(prizeStatistics);
-    const yieldRatio = this.getYieldRatio(totalPrizeMoney);
-    const prizeStatisticsTemplates = this.getPrizeStatisticsTemplates(prizeStatistics);
+    this.statisticsPhase(eachLottoPrize);
+  }
+
+  statisticsPhase(eachLottoPrize) {
+    this.statisticsMachine.makeStatisticsData(eachLottoPrize, this.purchaseAmount);
+    const prizeStatisticsTemplates = this.statisticsMachine.prizeStatisticsTemplates;
+    const yieldRatio = this.statisticsMachine.yieldRatio;
 
     this.LottoGameView.printPrizeStatistics(prizeStatisticsTemplates);
     this.LottoGameView.printYieldRatio(yieldRatio);
     Console.close();
   }
+
   getEachLottoNumbers() {
     return this.lottos.map((lotto) => lotto.getNumbers());
-  }
-
-  getPrizeStatistics(eachCalculatedLottoPrize) {
-    const prizeStatistics = {
-      fifthPlace: 0,
-      fourthPlace: 0,
-      thirdPlace: 0,
-      secondPlace: 0,
-      firstPlace: 0,
-      fail: 0,
-    };
-
-    eachCalculatedLottoPrize.forEach((lottoPrize) => (prizeStatistics[lottoPrize] += 1));
-
-    return prizeStatistics;
-  }
-  getTotalPrizeMoney(prizeStatistics) {
-    let totalPrizeMoney = 0;
-
-    for (const [prize, count] of Object.entries(prizeStatistics)) {
-      const prizeMoney = LOTTO_PRIZE_MONEY[prize] * count;
-      totalPrizeMoney += prizeMoney;
-    }
-
-    return totalPrizeMoney;
-  }
-  getYieldRatio(totalPrizeMoney) {
-    if (totalPrizeMoney) {
-      return ((totalPrizeMoney / this.purchaseAmount) * 100).toFixed(1);
-    }
-
-    return 0;
-  }
-  getPrizeStatisticsTemplates(prizeStatistics) {
-    const templates = [
-      `${
-        LOTTO_PRIZE_MATCH_COUNT.fifthPlace
-      }개 일치 (${LOTTO_PRIZE_MONEY.fifthPlace.toLocaleString()}원) - ${
-        prizeStatistics.fifthPlace
-      }개`,
-      `${
-        LOTTO_PRIZE_MATCH_COUNT.fourthPlace
-      }개 일치 (${LOTTO_PRIZE_MONEY.fourthPlace.toLocaleString()}원) - ${
-        prizeStatistics.fourthPlace
-      }개`,
-      `${
-        LOTTO_PRIZE_MATCH_COUNT.thirdPlace
-      }개 일치 (${LOTTO_PRIZE_MONEY.thirdPlace.toLocaleString()}원) - ${
-        prizeStatistics.thirdPlace
-      }개`,
-      `${
-        LOTTO_PRIZE_MATCH_COUNT.thirdPlace
-      }개 일치, 보너스 볼 일치 (${LOTTO_PRIZE_MONEY.secondPlace.toLocaleString()}원) - ${
-        prizeStatistics.secondPlace
-      }개`,
-      `${
-        LOTTO_PRIZE_MATCH_COUNT.firstPlace
-      }개 일치 (${LOTTO_PRIZE_MONEY.firstPlace.toLocaleString()}원) - ${
-        prizeStatistics.firstPlace
-      }개`,
-    ];
-
-    return templates;
   }
 }
